# [perf] Video writer optimize