This pr uses GL_PIXEL_PACK_BUFFER to accelerate the process of reading pixel data from gpu to cpu.
The pbo optimization can be toggled off in case of any incompatible situation
Also added a feature to automatically detect and select hardware encoder (Nvidia nvenc / AMD amf) for ffmpeg.

Performace inprovements:  Time cost of rendering 20 seconds of 60fps 1920x1080 video dropped from 13.53s to 8.34s on my i7-12700h / 3050Laptop device. (around 38% speed up)